Update and rename dij.py to dijkstra.py
